+++ src/components/form/FieldTextarea.vue
@@ -17,7 +17,6 @@ div(
     "dm-field-textarea",
     "dm-field-textarea--" + computedStatus,
     "dm-field-textarea--" + size,
-    "dm-field-textarea--" + theme,
     {
       "dm-field-textarea--borders": borders,
       "dm-field-textarea--disabled": disabled,